[ska][x-pack] Move saved_objects_management, saved_query_management and snapshot_restore to platform
(cherry picked from commit 917281648dbc379dbb18da18fc14a249825153ec)
