+++ x-pack/platform/test/functional/apps/saved_objects_management/config.ts
@@ -8,10 +8,10 @@
 import { FtrConfigProviderContext } from '@kbn/test';
 
 export default async function ({ readConfigFile }: FtrConfigProviderContext) {
-  const functionalConfig = await readConfigFile(require.resolve('../../config.base.js'));
+  const functionalConfig = await readConfigFile(require.resolve('../../config.base.ts'));
 
   return {
     ...functionalConfig.getAll(),
-    testFiles: [require.resolve('./feature_controls/v2')],
+    testFiles: [require.resolve('.')],
   };
 }
