+++ x-pack/platform/test/functional/apps/saved_objects_management/feature_controls/saved_objects_management_security.ts
