+++ x-pack/platform/test/functional/apps/snapshot_restore/home_page.ts
