+++ x-pack/platform/test/functional/apps/snapshot_restore/snapshot_restore.ts
