+++ x-pack/platform/test/functional/page_objects/index.ts
@@ -19,6 +19,7 @@ import { ReportingPageObject } from './reporting_page';
 import { WatcherPageObject } from './watcher_page';
 import { SearchProfilerPageProvider } from './search_profiler_page';
 import { UpgradeAssistantFlyoutObject } from './upgrade_assistant_page';
+import { SnapshotRestorePageProvider } from './snapshot_restore_page';
 
 // just like services, PageObjects are defined as a map of
 // names to Providers. Merge in Kibana's or pick specific ones
@@ -37,4 +38,5 @@ export const pageObjects = {
   watcher: WatcherPageObject,
   searchProfiler: SearchProfilerPageProvider,
   upgradeAssistant: UpgradeAssistantFlyoutObject,
+  snapshotRestore: SnapshotRestorePageProvider,
 };
