+++ x-pack/platform/test/functional/page_objects/snapshot_restore_page.ts
